+++ src/utils/device-utils.ts
@@ -53,13 +53,13 @@ export const supportsSetSinkId = (): boolean => {
   return 'setSinkId' in HTMLAudioElement.prototype;
 };
 
-export function isPrevNextObjectUndefined<T>(prev: T, next: T): boolean{
+export function isPrevNextUndefined<T>(prev: T, next: T): boolean {
   const isPrevUndefined = prev === undefined;
   const isNextUndefined = next === undefined;
   return isPrevUndefined && isNextUndefined;
 };
 
-export function isPrevNextObjectEmpty<T>(prev: T, next: T): boolean {
+export function isPrevNextEmpty<T>(prev: T, next: T): boolean {
   const isPrevEmpty = (prev && Object.keys(prev).length === 0);
   const isNextEmpty = (next && Object.keys(next).length === 0);
   return isPrevEmpty && isNextEmpty;